pyproject.toml
[project]
name = "bedrock-ge"
version = "0.3.2"
description = "Bedrock's Python library for geotechnical engineering."
authors = [
    {name = "Bedrock", email = "info@bedrock.engineer"}
]
license = {text = "Apache Software License (Apache 2.0)"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chardet>=5.2.0",
    "geopandas~=1.0",
    "openpyxl~=3.0",
    "pandera>=0.23.0",
    "python-ags4~=1.0",
    "sqlmodel>=0.0.22",
]
keywords = [
    "bedrock",
    "geotechnical-engineering",
    "ground-engineering",
    "ground-investigation",
    "ground-investigation-data",
    "gi-data",
    "ags",
    "ags3",
    "ags4",
    "civil-engineering",
    "aec",
    "aeco",
    "bim",
    "gis",
    "geospatial",
    "geospatial-data",
    "geo-bim",
    "geoscience-bim",
    "geotech",
    "geotechnical",
    "geotechnics",
    "engineering-geology",
    "geosciences",
    "geostatistics",
    "geotechnical-data",
    "borehole-data",
    "borehole",
    "subsurface",
    "underground"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: GIS",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
]

[project.urls]
Homepage = "https://bedrock.engineer/"
Source = "https://github.com/bedrock-engineer/bedrock-ge"
Documentation = "https://bedrock.engineer/docs/"
Tracker = "https://github.com/bedrock-engineer/bedrock-ge/issues"


[dependency-groups]
dev = [
    "frictionless[excel]>=4.40.8",
    "marimo[recommended]>=0.13.11",
    "mypy>=1.11.2",
    "pandas-stubs>=2.2.2.240807",
]

tests = [
    "folium>=0.17.0",
    "mapclassify>=2.8.1",
    "marimo>=0.13.11",
    "matplotlib>=3.9.2",
    "pytest>=8.3.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "D", "W"]
ignore = ["F401"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 88
